DataThresholdUI to specify a threshold over double[].
A UI component that displays the log10 histogram of a
double data array and let the user configure a threshold
over it.

The threshold is specified by 3 fields:
•the threshold value;
•whether the threshold is above or below this value;
•whether the comparison is strict.
The value currently set can be accessed by getThreshold().
It implements DoublePredicate and can used in a lambda
expression.
The JPanel that contains the UI is accessed via getPanel(),
 and relies on JFreeChart. The threshold is set by mouse
and keyboard. The arrow keys are used to move the threshold
value on the histogram. It is also possible to directly type
a value, like "16.34", wait, and this value will be grabbed.